code refactor, passing id instead of all gif to full screen, move navigation package above in project tree
